Initial commit.
"It's what you want to give to each person.
Your feelings, the feelings you believe they should have.
To give to the one you love her greatest dream.
To give her the ultimate emotion:
The feeling of total life, the vision of all the world's colors at once.
That is the reason for my existence."
-- Sharon Apple, Macross Plus